requirements.txt
Flask
langchain-core
langchain-community
langchain-openai
langchain-google-genai
google-genai
duckduckgo-search
newspaper3k
requests
beautifulsoup4
feedparser
wikipedia
yfinance